// front-end/packages/shared/src/models/orderStatus.ts
/**
 * Generated by Orval
 * Do not edit manually.
 * OpenAPI specification - RING-Bookstore
 */

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  PENDING_PAYMENT: "PENDING_PAYMENT",
  PENDING: "PENDING",
  SHIPPING: "SHIPPING",
  COMPLETED: "COMPLETED",
  CANCELED: "CANCELED",
  PENDING_RETURN: "PENDING_RETURN",
  PENDING_REFUND: "PENDING_REFUND",
  REFUNDED: "REFUNDED",
} as const;